[UPD]Refactoring project structures
